api/finance/models/user.py
from finance import db
from werkzeug.security import generate_password_hash, check_password_hash


class User(db.Model):
    """User

    The users that can access the system
    """

    pk = db.Column(db.Integer, primary_key=True)
    username = db.Column(db.String(50), unique=True)
    password_hash = db.Column(db.String(100))

    def __init__(self, username, password):
        self.username = username
        self.set_password(password)

    def set_password(self, password):
        self.password_hash = generate_password_hash(password)

    def check_password(self, password):
        return check_password_hash(self.password_hash, password)

    def __repr__(self):
        return '<User: {0}>'.format(self.username)
